[ROCm] run test_checkpoint.py only under default fa backend

+++ ci/pytorch.sh
@@ -41,10 +41,12 @@ run_default_fa_lbl() {
 run_test_config(){
     echo ==== Run with Fused attention backend: $_fus_attn ====
     #_WORKERS_COUNT=$TEST_WORKERS
-    mkdir -p ${TEST_DIR}/checkpoint
-    python ${TEST_DIR}/test_checkpoint.py --save-checkpoint all --checkpoint-dir ${TEST_DIR}/checkpoint
-    NVTE_TEST_CHECKPOINT_ARTIFACT_PATH=${TEST_DIR}/checkpoint run 1 test_checkpoint.py
-    rm -rf ${TEST_DIR}/checkpoint
+    if [ $_fus_attn = "$_DEFAULT_FUSED_ATTN" ]; then
+        mkdir -p ${TEST_DIR}/checkpoint
+        python ${TEST_DIR}/test_checkpoint.py --save-checkpoint all --checkpoint-dir ${TEST_DIR}/checkpoint
+        NVTE_TEST_CHECKPOINT_ARTIFACT_PATH=${TEST_DIR}/checkpoint run 1 test_checkpoint.py
+        rm -rf ${TEST_DIR}/checkpoint
+    fi
     run 1 test_cuda_graphs.py
     run_default_fa 1 test_deferred_init.py
     run_default_fa 1 test_float8tensor.py
